+++ PlaceRouteHierFlow/PnRPython/pytest.ini
@@ -1,0 +1,5 @@
+[pytest]
+addopts = --junitxml=junit.xml
+
+[pytest-watch]
+dont-watch-files = ^\.#
